Various mapping fixes

+++ BG3Extender/GameDefinitions/Skeleton.h
@@ -42,7 +42,7 @@ struct Skeleton : public ProtectedGameObject<Skeleton>
     Array<SkeletonSocket> Sockets;
     uint16_t BoneCount;
     std::array<uint16_t, 6> SoundBoneIndices;
-    std::array<uint16_t, 10> SoundActivationRanges;
+    std::array<float, 10> SoundActivationRanges;
     SkeletonContent* SkeletonContent;
 };
